refactor!: move pkg/kubediscovery to internal/kubediscovery
